// src/hunterxhunter/enums/tags.enum.ts
export enum TagsPtBr {}

export enum TagsEn {
  Abilities = '[data-source="abilities"] > div',
  AbilitiesDescription = 'table[class="wikitable"][cellpadding="2"][cellspacing="1/6"] > tbody > tr',
  Affiliation = '[data-source="affiliation"] > div|[data-source="previous affiliation"] > div',
  Age = '[data-source="age"] > div',
  Alias = '[data-source="kana"] > div|[data-source="rōmaji"] > div',
  Birthday = '[data-source="birthday"] > div > a > span',
  BloodType = '[data-source="blood"] > div',
  Description = '#mw-content-text > aside ~ p',
  EyeColor = '[data-source="eyes"] > div',
  Gender = '[data-source="gender"] > div',
  HairColor = '[data-source="hair"] > div',
  Height = '[data-source="height"] > div',
  JapaneseVoice = '[data-source="japanese voice"] > div',
  Name = '.page-header__title',
  NenType = '[data-source="type"] > div',
  Occupation = '[data-source="occupation"] > div|[data-source="previous occupation"] > div',
  Photo = '[data-source="image"] > div figure a|[data-source="image"] a',
  Relatives = '[data-source="relatives"] > div',
  Status = '[data-source="status"] > div',
  Weight = '[data-source="weight"] > div'
}

export enum TagsEs {
  Abilities = '[data-source="habilidades"] > div',
  AbilitiesDescription = 'table[class="wikitable"][cellpadding="2"][cellspacing="1/6"] > tbody > tr',
  Affiliation = '[data-source="afiliación"] > div',
  Age = '[data-source="edad"] > div',
  Alias = '[data-source="kana"] > div|[data-source="romaji"] > div',
  Birthday = '[data-source="cumpleaños"] > div',
  BloodType = '[data-source="tipo de sangre"] > div',
  Description = '#mw-content-text > aside ~ p',
  EyeColor = '[data-source="color de ojos"] > div',
  Gender = '[data-source="genero"] > div',
  HairColor = '[data-source="color de pelo"] > div',
  Height = '[data-source="altura"] > div',
  JapaneseVoice = '[data-source="japones"] > div',
  Name = '.page-header__title',
  NenType = '[data-source="tipo"] > div',
  Occupation = '[data-source="ocupacion"] > div',
  Photo = '[data-source="image"] > div figure a|[data-source="image"] a',
  Relatives = '[data-source="familia"] > div',
  Status = '[data-source="status"] > div',
  Weight = '[data-source="peso"] > div'
}